refactored pet-form using zod

+++ src/actions/actions.ts
@@ -1,51 +1,58 @@
 'use server';
 import prisma from '@/lib/db';
-import { PetEssentials } from '@/lib/types';
 import { sleep } from '@/lib/utils';
-import { Pet } from '@prisma/client';
+import { petFormSchema, petIdSchema } from '@/lib/validations';
 import { revalidatePath } from 'next/cache';
 
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
   await sleep(1000);
+
+  const validatedPet = petFormSchema.safeParse(pet);
+
+  if (!validatedPet.success) return { message: 'Invalid pet data' };
 
   try {
     await prisma.pet.create({
-      data: pet,
+      data: validatedPet.data,
     });
   } catch (error) {
-    return {
-      message: "Couldn't add pet",
-    };
+    return { message: "Couldn't add pet" };
   }
 
   revalidatePath('/app', 'layout');
 }
 
-export async function editPet(petId: Pet['id'], newPetData: PetEssentials) {
+export async function editPet(petId: unknown, newPetData: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(newPetData);
+
+  if (!validatedPetId.success || !validatedPet.success) {
+    return { message: 'Invalid pet data' };
+  }
 
   try {
     await prisma.pet.update({
-      where: {
-        id: petId,
-      },
-      data: newPetData,
+      where: { id: validatedPetId.data },
+      data: validatedPet.data,
     });
   } catch (error) {
-    return {
-      message: "Couldn't edit pet",
-    };
+    return { message: "Couldn't edit pet" };
   }
 
   revalidatePath('/app', 'layout');
 }
 
-export async function deletePet(petId: Pet['id']) {
+export async function deletePet(petId: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+  if (!validatedPetId.success) return { message: 'Invalid pet ID' };
 
   try {
     await prisma.pet.delete({
-      where: { id: petId },
+      where: { id: validatedPetId.data },
     });
   } catch (error) {
     return { message: "Couldn't delete pet" };
